# ss/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movies-app-ss",
        "version": "1.0.0"
    },
    "paths": {
        "api/ads": {
            "get": {
                "tags": [
                    "Ad"
                ],
                "summary": "Get list of ads.",
                "operationId": "2ac83226d75045908f54059aab7a2c3b",
                "responses": {
                    "200": {
                        "description": "Display a listing of ads."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ad"
                ],
                "summary": "Create new ad.",
                "operationId": "98cf29c88c33fbcbf8ae29145d77d599",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of ad",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/name"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image of ad.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/image"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ad"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Create successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "api/ads/{id}": {
            "get": {
                "tags": [
                    "Ad"
                ],
                "summary": "Get ad by id",
                "operationId": "4c02eadec1ea36728db4a2b4a3627e51",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display a listing of films."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ad"
                ],
                "summary": "Update ad by id.",
                "operationId": "ad07341613dcd1f42fee351a4d407151",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/id"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of ad",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/name"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Image of ad.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/image"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Ad"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ad"
                ],
                "summary": "Delete ad by id.",
                "operationId": "a13dead5f40f2ab3376c587219c3b83c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Ad/properties/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete successfully."
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Ad": {
                "required": [
                    "name",
                    "image"
                ],
                "properties": {
                    "id": {
                        "$ref": "#/components/schemas/BaseModel/properties/id"
                    },
                    "name": {
                        "type": "string",
                        "example": "Ad 1"
                    },
                    "image": {
                        "type": "string",
                        "example": "upload/image.jpg"
                    },
                    "status": {
                        "type": "integer",
                        "example": "Trạng thái: 1-kích hoạt, 2-vô hiệu, 3-đã xóa"
                    },
                    "position": {
                        "type": "integer",
                        "example": "1"
                    },
                    "created_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_by"
                    },
                    "updated_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_by": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "update_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    },
                    "deleted_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/deleted_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Ad"
                }
            },
...
        }
    }
}